pyproject.toml
[tool.poetry]
name = "eth-validator-watcher"
version = "0.1.0"
description = "Ethereum Validator Watcher"
authors = ["Manu NALEPA <emmanuel.nalepa@kiln.fi>"]
readme = "README.md"
packages = [{include = "eth_validator_watcher"}]

[tool.poetry.dependencies]
python = "^3.9"
more-itertools = "^9.1.0"
prometheus-client = "^0.16.0"
pydantic = "^1.10.7"
requests = "^2.28.2"
sseclient-py = "^1.7.2"
typer = "^0.7.0"
slack-sdk = "^3.21.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
requests-mock = "^1.10.0"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
eth-validator-watcher = "eth_validator_watcher.entrypoint:app"